# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.ContentType:
    enum:
    - text
    - emoji
    - image
    - link
    type: string
    x-enum-comments:
      Emoji: 表情
      Image: 图片
      Link: 链接
      Text: 文本
    x-enum-descriptions:
    - 文本
    - 表情
    - 图片
    - 链接
    x-enum-varnames:
    - Text
    - Emoji
    - Image
    - Link
...
  model.LoginRequest:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - password
    - username
    type: object
...
  model.MessagePart:
    properties:
      content:
        description: 内容（如文本、图片 URL、链接等）
        type: string
      type:
        allOf:
        - $ref: '#/definitions/model.ContentType'
        description: 内容类型（text, emoji, image, link）
    type: object
  model.MessageType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-comments:
      ForwardedCotent: 转发
      ImageContent: 图片
      RedBagContent: 红包
      SystemContent: 系统消息
      TextContent: 聊天消息
      VoiceContent: 语音
    x-enum-descriptions:
    - 聊天消息
    - 图片
    - 语音
    - 红包
    - 转发
    - 系统消息
    x-enum-varnames:
    - TextContent
    - ImageContent
    - VoiceContent
    - RedBagContent
    - ForwardedCotent
    - SystemContent
  model.MessageVo:
    properties:
      content:
        items:
          $ref: '#/definitions/model.MessagePart'
        type: array
      createdAt:
        type: string
      extra_data:
        description: 扩展字段（如红包、投票等结构）
      groupId:
        format: int64
        type: integer
      id:
        type: integer
      isRead:
        type: boolean
      readerIdList:
        items:
          type: integer
        type: array
      receiverId:
        format: int64
        type: integer
      reply:
        allOf:
        - $ref: '#/definitions/model.MessageVo'
        description: 额外信息
      replyId:
        format: int64
        type: integer
      senderAvatar:
        type: string
      senderId:
        format: int64
        type: integer
      senderNickName:
        type: string
      senderOnlineStatus:
        $ref: '#/definitions/model.OnlineStatus'
      status:
        $ref: '#/definitions/model.Status'
      targetType:
        $ref: '#/definitions/model.TargetType'
      type:
        $ref: '#/definitions/model.MessageType'
      updatedAt:
        type: string
    type: object
  model.OnlineStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      Away: 3 离开
      Busy: 2 忙碌
      Offline: 0 离线
      Online: 1 在线
    x-enum-descriptions:
    - 0 离线
    - 1 在线
    - 2 忙碌
    - 3 离开
    x-enum-varnames:
    - Offline
    - Online
    - Busy
    - Away
  model.QueryMessagesRequest:
    properties:
      cursor:
        description: 游标 上次查询的最小id 默认为0
        type: integer
      end_time:
        description: 结束时间
        type: string
      keyword:
        description: 模糊搜索
        type: string
      limit:
        description: 限制数量
        type: integer
      message_types:
        description: 以下为拓展,暂时不会做
        items:
          $ref: '#/definitions/model.MessageType'
        type: array
      start_time:
        description: 起始时间
        type: string
      target_id:
        description: 目标ID 好友id或群组id
        type: integer
      target_type:
        allOf:
        - $ref: '#/definitions/model.TargetType'
        description: 目标类型 私聊或群聊
    required:
    - target_id
    - target_type
    type: object
  model.QueryMessagesResponse:
    properties:
      cursor:
        description: 下一页游标（最小 id）
        type: integer
      has_more:
        description: 是否还有更多数据
        type: boolean
      list:
        description: 消息列表
        items:
          $ref: '#/definitions/model.MessageVo'
        type: array
    type: object
  model.ReadMessageReq:
    properties:
      message_id:
        type: integer
      user_id:
        type: integer
    type: object
  model.Response:
    description: 通用响应格式，用于包装所有 API 响应
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据部分
      message:
        description: 提示信息
        type: string
    type: object
...
  model.Status:
    enum:
    - 0
    - 1
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      Accept: 1 接受
      Disable: 0 禁用
      Enable: 1 启用
      Reject: 2 拒绝
      Todo: 0 待处理
    x-enum-descriptions:
    - 0 禁用
    - 1 启用
    - 0 待处理
    - 1 接受
    - 2 拒绝
    x-enum-varnames:
    - Disable
    - Enable
    - Todo
    - Accept
    - Reject
  model.TargetType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-comments:
      GroupTarget: 群聊
      PrivateTarget: 私聊
    x-enum-descriptions:
    - 私聊
    - 群聊
    x-enum-varnames:
    - PrivateTarget
    - GroupTarget
...
  model.UserUpdateRequest:
    properties:
      avatar:
        type: string
      desc:
        maxLength: 500
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        maxLength: 60
        type: string
      phone:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /message/{id}/revoke:
    get:
      consumes:
      - application/json
      description: 根据消息ID撤回指定消息，只有发送者或管理员才能撤回消息
      parameters:
      - description: 消息ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功"  // 成功的撤回响应
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 撤回失败"  // 撤回消息失败
          schema:
            $ref: '#/definitions/model.Response'
      summary: 撤回消息
      tags:
      - Message
  /message/query:
    post:
      consumes:
      - application/json
      description: 根据目标ID和目标类型查询聊天消息历史，支持分页、时间范围等过滤
      parameters:
      - description: 查询参数
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/model.QueryMessagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回消息列表及分页信息
          schema:
            $ref: '#/definitions/model.QueryMessagesResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 查询历史消息（分页，支持游标分页）
      tags:
      - Message
  /message/read:
    post:
      consumes:
      - application/json
      description: 对该条消息已读
      parameters:
      - description: 消息内容
        in: body
        name: msg
        required: true
        schema:
          $ref: '#/definitions/model.ReadMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 发送消息失败
          schema:
            $ref: '#/definitions/model.Response'
      summary: 已读消息
      tags:
      - Message
  /user/info:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功"  // 返回当前用户的信息
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 获取用户信息失败"  // 获取用户信息失败
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取用户信息
      tags:
      - user
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录，返回用户的认证信息
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: 登陆失败
          schema:
            $ref: '#/definitions/model.Response'
      summary: 用户登录
      tags:
      - user
  /user/logout:
    get:
      consumes:
      - application/json
      description: 用户登出，清除认证信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功"  // 成功的登出响应
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 登出失败"  // 登出失败
          schema:
            $ref: '#/definitions/model.Response'
      summary: 用户登出
      tags:
      - user
  /user/online_status_change:
    get:
      consumes:
      - application/json
      description: 修改用户的在线状态（例如：在线、离线、忙碌等）
      parameters:
      - description: 在线状态
        in: query
        name: online_status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功"  // 在线状态修改成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误"  // 请求参数错误
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: 在线状态变更失败"  // 状态变更失败
          schema:
            $ref: '#/definitions/model.Response'
      summary: 用户在线状态变更
      tags:
      - user
  /user/register:
    post:
      consumes:
      - application/json
      description: 接口用于用户注册，提供注册所需的字段
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.Response'
      summary: 用户注册
      tags:
      - user
  /user/update:
    post:
      consumes:
      - application/json
      description: 更新用户的个人资料，如昵称、头像等
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功"  // 更新成功的响应
          schema:
            type: string
        "400":
          description: 请求参数错误"  // 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权"  // 未授权，必须登录后更新
          schema:
            type: string
        "500":
          description: 更新失败"  // 更新失败
          schema:
            type: string
      security:
      - Bearer: []
      summary: 更新用户信息
      tags:
      - user
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"